ENH: First(?) good practice recommendation. No excessive overrides in Inheritance principle
It is often useful to look at the top level task*json files to see what is
"common" among all subjects/sessions.  Listing only the fields with common
values is what we also implemented (albeit incorrectly, fix is pending
https://github.com/nipy/heudiconv/pull/279) in ReproIn.

Upon quick search I also found no "good practice" recommendations in the
spec.  It might be arguable either spec should or should not include them.
I think it should since we already RECOMMEND things.  We of cause could
place recommendations into a separate subsection somewhere but then they
would not be linked to the specific topics and thus would not be attended
to. Hence I placed this one right at the section where it is pertinent