Added AcquisitionArea as metadata field
The field AcquisitionArea can be used to indicate what can be seen in the corresponding files, e.g. brain or parts of the brain, spinal cord, peripheral nerves, etc.
Guidelines for reporting fMRI studies (Poldrack et al 2008) also recommend using descriptions of the acquisition area when the images do not encompass the whole brain, the field AcquisitionAreaDescription can be used for that.