[ENC] getting rid of type- key
I don't know where the type key came from. 
I also tried to remove the reference to a space table in the introduction that has moved (is it in the appendix?)

(I'll post this an issue)
So currently we have. _dseg.tsv suffix for color lookup tables. A label-<label> key value pair for specifying the label in _probseg.nii.gz files. And files with multiple mask (i.e. atlases having a _dseg.nii.gz) extension. And and atlas-<label> key value pair that appears in fmri..

1. is it ok for _dseg.nii.gz and _dseg.tsc to have the same suffix? or are suffixes supposed to associate with a specific file type.
2. technically the _dparc.dlabel.nii could be _dseg.dlabel.nii because they have similar concepts (just cifti vs nifti).
3.