enh: add "multipart DWI" acquisitions and refactor DWI specifications
Addresses the use case of DWI sequencies that need to be split in several
runs because, e.g., the scanner would exceed temperature specifications
due to fast gradient-switching.

This PR also refactors the text of DWIs to better accomodate the addition
and make DWIs more consistent with the rest of the file.

The approach to solve this is similar to that of #622 for fieldmaps.

References: nipreps/dmriprep#43.