enh: allow encoding the fieldmapping intent of the protocol
This PR addresses the problem @mattcieslak spotted at in #239.

This enhancement (WIP) basically allows for researchers to encode
the protocol's intent regarding fieldmaps.

As @satra introduced in #239 (comment),
BIDS "*could encode intent and automation. Whether it should is a community decision."

This PR proposes a solution to encoding the intent. It doesn't modify
anything to allow also encoding automation.

The PR attempts to be backwards compatible, and is based off of #651,
where the text about fieldmaps is being revised.

I'm submitting this draft PR to open discussions and looking forward to
feedback.

Resolves: #239.
Depends: #651.
References: #263, nipreps/dmriprep#43, bids-standard/bids-2-devel#39