Inheritance principle: Relaxation allowing multiple files per directory
Relates to discussion in #259, and was also raised in #946.
Instead of enforcing a unique order of metadata file inheritance via filesystem hierarchy only, the logic behind the inheritance principle is generalised to permit multiple files per directory whilst still ensuring that the order of metadata file loading is still unique.