ENH: introduce versioning of the schema itself (#1013)
* ENH: introduce versioning of the schema itself

Note -- it is not a version of the content of the schema, it is
a version over organization of the schema (e.g. we could add some
schema to describe how schema is organized).

The 'desire' for having a version over the BIDS schema's version
is somewhat described in README.md in added section.  Just to facilitate
tools to know what versions of bids schema they can read without explicitly
whitelisting BIDS versions themselves.

* Adding versioning for after 1.0.0

I thought we had it but managed to screw it up during rebase. Thanks @sappelhoff

Co-authored-by: Stefan Appelhoff <stefan.appelhoff@mailbox.org>

* line breaks

Co-authored-by: Stefan Appelhoff <stefan.appelhoff@mailbox.org>