ENH: use __file__ instead of getcwd to determine HERE, provide informative message on # of files (#1091)
I had to spend some time to figure out why running that script did nothing for me
but failed in CI for the PR.  Apparently it was because I had to run it while
in tools/ directory.  IMHO it is suboptimal.

With a short message on how many files were analyzed user at least could
get a hint in case if no files were read at all

HERE is no longer used, and thuse removed entirely.