ENH: Adds mdp suffix.
Following up on #1280, this PR differentiates between model parameters
and model-derived parameters. This distinction arises from considerations
described in more detail in BEP16. For example, the distinction
between the 6 elements of the tensor that are fit in the process of
DTI, and secondary calculations of FA and MD from the eigenvalues
of the tensor.