Add to schemacode_ci.yml a workflow to publish master version of schema
Unfortunately we cannot use this workflow for publishing on releases since
it reacts to schema- tags and not v* tags and anyways not taking care about
actual releases of the bids-specification itself.  So we would either need to
move this into release workflow (is it the circleci one?) or just duplicate this
directly in  bids-schema  making it to run there on cron and checking version tags.