Remove explicitly worded requirement of events file association with the task
The discussion came up during reviewing BEP032 that association of
events file with an active agent performing a task might be of a stretch
for many animal experiments.

Other arguments
- we used _events suffix, not _taskevents  so to me it makes total sense
  to generalize _events to all events
- wording already suggests that events could be stimuli and participant
  responses thus offering support that events could be related to both
  influences over participant and behavioral responses thus mixing two
  domains of data. Metadata fields such as event_type remains there to
  possibly disentangle multiple domains

As a result we would open possibility to make record of events during
non-functional runs, e.g.

- someone interested to investigate effect of various "events" which occured
  during acquisition of anatomical data (which can spend minutes), e.g.
  "change of screen intensity", "instruction input", etc, which could later
  be used e.g. for removal of confounds etc.

This is an incomplete change to first spark discussion and further
improvements