Add ContextURI to allow to define the context for the entity values
It is often desired to be able to determine that values used
for entities in the dataset belong to some controlled vocabulary,
or simply defined centrally within some "id" authority. E.g. could be
unique scanning session IDs per scanning center, or similarly subject_ids defined
per study or centrally for the center.

It is of particular interest for large studies where multiple datasets could be
created, one per site or primary data modality, to later possibly be composed
into a single dataset or just to become parts of the one larger multi-site one.
In such cases it becomes quite important to annotate that particular entities
(subject_id, session_id and possibly even _desc- or _acq- values) are defined
in the scope of the specific  larger study and thus correspond to the "same" thing
given the same contextURI and value.