fix: avoid using an atlas name for a segmentation label
As @CPernet mentioned, the spec was unclear because the example illustrating
``seg-`` was using an atlas name for the label.

I grepped the whole spec for ``seg-`` and only found ``seg-Desikan``.

This commit replaces that label, edits a bit the text around it and
updates the `seg-` entity definition in the schema to be more clear.

Even though its use with ``atlas-`` was already mentioned:

> For atlases, `seg-<label>` distinguish different realizations of a given
> segmentation or parcellation.
> For example, the [Yeo 2011 atlas](https://doi.org/10.1152/jn.00338.2011)
> is distributed within *FreeSurfer* with two different parcellations
> (*7 networks* and *17 networks*).

I have updated the entity's description to be more clear (hopefully).