RF(schema): harmonize into dirs, ids, ids_phenotype
Current names duplicate their "domain" (subject_ or session_), inconsistent in
plurality (_id although for all ids), and not clear really what
'phenotype' corresponds to without reading the description.

With proposed change there is no duplication of the domain, consistency
in plurality (albeit 'id' is an abbreviation so 'ids' is a non-word), and
IMHO clearer meaning in `ids_phenotype`.

This would also allow for generalization across other entities in a perspective
https://github.com/bids-standard/bids-2-devel/issues/54 - where then any entity
with folders for its level could have `dirs`. Also it could come handy to
determine `ids` for some other entities in tests.

Ref: https://github.com/bids-standard/bids-validator/pull/94#discussion_r1836876737

TODOs
- [ ] introduce corresponding changes to bids-validator