Allow `acq-*` to relate physio data to MRI data in `anat/` as described in #2164
The deno based validator did not accept files like `sub-01_acq-T1w_physio.tsv.gz` located in `anat/` and related to the T1w anatomy. This was, because a task was required, which was inherited from `rules.files.raw.task.timeseries.entities`. The key task is now overridden to optional. Please also see issue #2164 for more information and related patches.