Do explicitly "allow" for having dotfiles and explicitly ignore them
Up to this point it was up for a validator to "code" that up.
E.g. current new deno based validator in
https://github.com/bids-standard/bids-validator/blob/main/src/files/ignore.ts
has

    const defaultIgnores = [
      '.git**',
      '.*',
      'sourcedata/',
      ...

thus ignores all dotfiles from validation.  But this is inconsistent e.g.  with
shipped in bids-specification schematools treatment, where files in a
dotdirectory were reported as "invalid":

	❯ cat /tmp/input.txt | python -m bidsschematools pre-receive-hook
	.datalad/config
	.gitattributes
	sub-01/anat/sub-01_unknown.nii.gz

And as demonstrated above there could be legit "system" dotfiles (VCS related
etc) which are part of the folder containing BIDS dataset, I think we should
**explicitly** describe position of BIDS in relation to the dotfiles.  As it is
pretty much "ignore", that is what this PR is intended to state.

In this PR I also adjusted `pre_receive_hook` to ignore dotfiles.
The `find_files` was already ignoring them.