Do check for being a pull request not at job level
Since this way it keeps it "grey" on PRs, which causes all other
matrix runs and workflows to be expanded and thus requiring more
mental cycles to understand that it is actually all good.

This way it should be collapsed.  This commit sits on top of
https://github.com/bids-standard/bids-specification/pull/2030
to ensure no other workflows fail ATM