Event onsets MUST NOT be negative
I found not a single openneuro datasets with negative onsets.

    smaug:/mnt/datasets/datalad/crawl/openneuro
    *$> for ds in ds*; do git grep '^ *-' '**_events.tsv' | head ; done

But ATM we are looking at a dataset with some buggy _events.tsv which has huge
negative values sorted within onsets column.  I think in general/theory in some
.1% of the cases there could be point to have onsets with negative onsets, e.g.
due to desire to encode some preparatory to the data scanning activities  (e.g.
during dummy scans etc).  So for those -- it would be nice to have a dedicated
ERROR code to ignore/overwrite.  But in general, it just signals a problem with
the file and thus IMHO should be an ERROR.