Revert "FIX: update PET examples"
This reverts commit 891bcf01bd9638e37909302c02c314e5648a3c27.

The changes in that commit were made before our latter meetings by the
end of 2024, and IMHO do not align with what we discussed.
However, @mnoergaard's suggestions did address a specific issue we
discussed in those meetings.
For that reason, I'm keeping the commit in the history so that the specific
suggestions are not lost, and I formally commit here to revisit the
issue and add examples with Martin as co-author, re-including his
suggestion.