Add .nwb as supported format for EEG
NWB is BRAIN Initiative supported, and open data standard
https://neurodatawithoutborders.github.io/ .

The main motivation here is to support joint iEEG and EEG datasets for
EMBER archive project.  Quick google can show that there is a good number of
studies which have conjoint EEG and iEEG.  Since .nwb is supported for iEEG, it
makes only sense to add support for it in EEG modality as well.

There is a good number of .nwb "specific" tools in the ecosystem, with a
notable mention of https://neurosift.app which allows "smart" client-side (in
browser) access to remote .nwb files for exploration.

It seems that there is also growing number of toolkits in EEG world which
support NWB.  E.g., EEGLAB already input support for NWB:
https://eeglab.org/plugins/NWB-io/ so data would be accessible in this popular
toolkit.

Added listing of NWB into EEG formats table + harmonize (change URL also)
in iEEG section