Remove reference to dicm2nii altogether
links checker started to fail, and I see

    ❯ wget -S https://www.mathworks.com/matlabcentral/fileexchange/42997-xiangruili-dicm2nii
    --2025-05-21 16:18:06--  https://www.mathworks.com/matlabcentral/fileexchange/42997-xiangruili-dicm2nii
    Resolving www.mathworks.com (www.mathworks.com)... 23.197.176.176
    Connecting to www.mathworks.com (www.mathworks.com)|23.197.176.176|:443... connected.
    HTTP request sent, awaiting response...
      HTTP/1.1 503 Service Unavailable

and in the browser I see "no healthy upstream" for that URL.

https://github.com/xiangruili/dicm2nii was indeed last updated 9 months ago and
overall IMHO just a very basic (no unittests etc) project.  I think
standard should not promote/point to tools which are not developed with active
participation in the project. dcm2niix is actively developed in sync and
participation in BIDS, so IMHO worth leaving in.