Add 1PE to complement 2PE
https://blog.biodock.ai/one-vs-two-photon-microscopy/ seems to provide a nice comparison
between one- and two-photon microscopy. Also
https://www.gophotonics.com/community/what-are-single-photon-and-two-photon-excitation
provides another comparison and uses "excitation" consistently between them.

Note that there is also already "FLUO" suffix for "Fluorescence microscopy" in general,
of which both 1PE and 2PE and likely MPE (multi-photon) mechanisms are, so unclear when
FLUO to be used, but I guess 2PE, MPE and now 1PE should be RECOMMENDED to be used
in favor of FLUO as more specific