[SCHEMA] Add "directories" description for "study" DatasetType and add "docs" where was missing (#2185)
* Add "directories" description for "study" DatasetType and add "docs" where was missing

Apparently I have missed this file entirely whenever I was preparing
  https://github.com/bids-standard/bids-specification/pull/1972 (study DatasetType)

and also due to all the duplication we (I) missed that docs was not listed among "root.subdirs"
or not listed at all.

This would provide a fix, but I wonder if we could/should make it so we could avoid duplication
altogether.  As I have argued in #1972 I feel that "study" is the base dataset type and
next ones just add potentially more to them.  So may be we could come up with some more
compact representation here... but not in this PR

* Remove "phentoype/" and "stimuli/" from the "root" of the "study" dataset

Since per se they should either be under "sourcedata/" or "derivatives/" one way or
another even if it is some stimuli-only dataset(s)