Update current BIDS maintainers list
Hi all,

with new priorities that came up for me, I have over the past year not been having enough time to make meaningful contributions to BIDS as a maintainer. Therefore, I would like to move my name from active to past BIDS maintainers.

That is the only change though -- as before I will still be available and involved in the project when needed, and/or when I find capacity to pick up tasks.