Boost dataset_description up in search results and changelog and glossary down
Immediate usecase came with @asmacdo finding GeneratedBy in glossary but without
connection to datataset_description.json and thus not being certain if that is
the location for it.  Overall, I think that glossary and changelog (even lower) is good
to search (so we must not exclude) but should be closer to the bottom of the list of the
hits since they would take longer to load, and would be "too generic" of hits.

Lowering changelog below glossary though did not bring desired effect for me and
changelog comes still above glossary when searching for GeneratedBy.